requirements.txt
aiokafka
fastapi
python-dotenv
uvicorn